% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efbreathr.R
\name{expfactory_breath_counting_to_csv}
\alias{expfactory_breath_counting_to_csv}
\title{Convert Experiment Factory breath counting data to CSV}
\usage{
expfactory_breath_counting_to_csv(path, t, exclude)
}
\arguments{
\item{path}{Path to data files}

\item{t}{Time point}

\item{exclude}{Participants to exclude}
}
\description{
Convert Experiment Factory breath counting data to CSV
}
