% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efbreathr.R
\name{process_breath_counting}
\alias{process_breath_counting}
\title{Process Breath Counting Task data file}
\usage{
process_breath_counting(path, p, json = TRUE)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}

\item{json=TRUE}{Boolean indicating whether ANT data format is JSON (TRUE) or CSV (FALSE)}
}
\value{
Data frame
}
\description{
\code{process_ant()} processes data files generated by the Breath Counting Task
(\url{https://expfactory.github.io/breath_counting_task.html}).
}
\examples{
ant <- expand.grid(token = participants$token) \%>\%
 rowwise() \%>\%
 mutate(path=paste0('data/', token, '_finished/breath-counting-task-results.json')) \%>\%
 do(., efbcr::process_breath_counting(.$path, .$token)) \%>\%

}
\keyword{breath}
\keyword{expfactory}
\keyword{mindfulness}
