% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efbreathr.R
\name{process_eprime_file}
\alias{process_eprime_file}
\title{Process ePrime Breath Counting Data File}
\usage{
process_eprime_file(path)
}
\arguments{
\item{path}{Path to ePrime data file}
}
\description{
Process ePrime breath counting data file
(Levinson, Stoll, Kindy, Merry, & Davidson, 2014)
}
\keyword{ePrime}
