% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efbreathr.R
\name{summarise_breath_counting_accuracy}
\alias{summarise_breath_counting_accuracy}
\title{Summarise breath counting accuracy}
\usage{
summarise_breath_counting_accuracy(data, participants)
}
\arguments{
\item{data}{data.frame returned by breath_counting_accuracy()}

\item{participants}{data.frame with column 'p' listing participants to process}
}
\value{
data.frame
}
\description{
Summarise breath counting accuracy
}
